+++ src/app/@theme/pipes/index.ts
@@ -3,4 +3,3 @@ export * from './plural.pipe';
 export * from './round.pipe';
 export * from './timing.pipe';
 export * from './number-with-commas.pipe';
-export * from './eva-icons.pipe';