+++ src/app/pages/editors/editors.module.ts
@@ -1,4 +1,5 @@
 import { NgModule } from '@angular/core';
+import { NbCardModule } from '@nebular/theme';
 import { CKEditorModule } from 'ng2-ckeditor';
 
 import { ThemeModule } from '../../@theme/theme.module';
@@ -7,6 +8,7 @@ import { EditorsRoutingModule, routedComponents } from './editors-routing.module
 
 @NgModule({
   imports: [
+    NbCardModule,
     ThemeModule,
     EditorsRoutingModule,
     CKEditorModule,