+++ src/app/pages/maps/maps.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { AgmCoreModule } from '@agm/core';
 import { LeafletModule } from '@asymmetrik/ngx-leaflet';
 import { NgxEchartsModule } from 'ngx-echarts';
+import { NbCardModule } from '@nebular/theme';
 
 import { ThemeModule } from '../../@theme/theme.module';
 import { MapsRoutingModule, routedComponents } from './maps-routing.module';
@@ -16,6 +17,7 @@ import { MapsRoutingModule, routedComponents } from './maps-routing.module';
     LeafletModule.forRoot(),
     MapsRoutingModule,
     NgxEchartsModule,
+    NbCardModule,
   ],
   exports: [],
   declarations: [