+++ src/app/pages/profile/profile.module.ts
@@ -1,12 +1,14 @@
 import { NgModule } from '@angular/core';
 import { ThemeModule } from '../../@theme/theme.module';
 import { ExtraComponentsModule } from '../extra-components/extra-components.module';
+import { NbCardModule } from '@nebular/theme';
 import * as components from './index';
 
 @NgModule({
   imports: [
     ThemeModule,
     ExtraComponentsModule,
+    NbCardModule,
   ],
   declarations: [
     components.ProfileComponent,